# backend/core/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://walletgpt.info:5050"
    }
  ],
  "paths": {
...
    "/brains/": {
      "get": {
        "tags": [
          "Brain"
        ],
        "summary": "Brain Endpoint",
        "description": "Retrieve all brains for the current user.\n\n- `current_user`: The current authenticated user.\n- Returns a list of all brains registered for the user.\n\nThis endpoint retrieves all the brains associated with the current authenticated user. It returns a list of brains objects\ncontaining the brain ID and brain name for each brain.",
        "operationId": "brain_endpoint_brains__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Brain"
        ],
        "summary": "Create Brain Endpoint",
        "description": "Create a new brain with given\n    name\n    status\n    model\n    max_tokens\n    temperature\nIn the brains table & in the brains_users table and put the creator user as 'Owner'",
        "operationId": "create_brain_endpoint_brains__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrainProperties"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      }
    },
    "/brains/default/": {
      "get": {
        "tags": [
          "Brain"
        ],
        "summary": "Get Default Brain Endpoint",
        "description": "Retrieve the default brain for the current user. If the user doesnt have one, it creates one.\n\n- `current_user`: The current authenticated user.\n- Returns the default brain for the user.\n\nThis endpoint retrieves the default brain associated with the current authenticated user.\nThe default brain is defined as the brain marked as default in the brains_users table.",
        "operationId": "get_default_brain_endpoint_brains_default__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      }
    },
    "/brains/{brain_id}/": {
      "get": {
        "tags": [
          "Brain"
        ],
        "summary": "Get Brain Endpoint",
        "description": "Retrieve details of a specific brain by brain ID.\n\n- `brain_id`: The ID of the brain to retrieve details for.\n- Returns the brain ID and its history.\n\nThis endpoint retrieves the details of a specific brain identified by the provided brain ID. It returns the brain ID and its\nhistory, which includes the brain messages exchanged in the brain.",
        "operationId": "get_brain_endpoint_brains__brain_id___get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Brain Id",
              "type": "string",
              "format": "uuid"
            },
            "name": "brain_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Brain"
        ],
        "summary": "Update Brain Endpoint",
        "description": "Update an existing brain with new brain configuration",
        "operationId": "update_brain_endpoint_brains__brain_id___put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Brain Id",
              "type": "string",
              "format": "uuid"
            },
            "name": "brain_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrainUpdatableProperties"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      }
    },
    "/brains/{brain_id}/default": {
      "post": {
        "tags": [
          "Brain"
        ],
        "summary": "Set As Default Brain Endpoint",
        "description": "Set a brain as default for the current user.",
        "operationId": "set_as_default_brain_endpoint_brains__brain_id__default_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Brain Id",
              "type": "string",
              "format": "uuid"
            },
            "name": "brain_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      }
    },
...
    "/explore/": {
      "get": {
        "tags": [
          "Explore"
        ],
        "summary": "Explore Endpoint",
        "description": "Retrieve and explore unique user data vectors.",
        "operationId": "explore_endpoint_explore__get",
        "parameters": [
          {
            "description": "The ID of the brain",
            "required": true,
            "schema": {
              "title": "Brain Id",
              "type": "string",
              "description": "The ID of the brain",
              "format": "uuid"
            },
            "name": "brain_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          }
        ]
      }
    },
    "/explore/{file_name}/": {
      "get": {
        "tags": [
          "Explore"
        ],
        "summary": "Download Endpoint",
        "description": "Download a specific user file by file name.",
        "operationId": "download_endpoint_explore__file_name___get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "File Name",
              "type": "string"
            },
            "name": "file_name",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Explore"
        ],
        "summary": "Delete Endpoint",
        "description": "Delete a specific user file by file name.",
        "operationId": "delete_endpoint_explore__file_name___delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "File Name",
              "type": "string"
            },
            "name": "file_name",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Brain Id",
              "type": "string",
              "format": "uuid"
            },
            "name": "brain_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthBearer": []
          },
          {
            "AuthBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrainUpdatableProperties": {
        "title": "BrainUpdatableProperties",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "openai_api_key": {
            "title": "Openai Api Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "prompt_id": {
            "title": "Prompt Id",
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateBrainProperties": {
        "title": "CreateBrainProperties",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "default": "Default brain"
          },
          "description": {
            "title": "Description",
            "type": "string",
            "default": "This is a description"
          },
          "status": {
            "title": "Status",
            "type": "string",
            "default": "private"
          },
          "model": {
            "title": "Model",
            "type": "string",
            "default": "gpt-3.5-turbo"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number",
            "default": 0
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer",
            "default": 256
          },
          "openai_api_key": {
            "title": "Openai Api Key",
            "type": "string"
          },
          "prompt_id": {
            "title": "Prompt Id",
            "type": "string",
            "format": "uuid"
          },
          "linkedin": {
            "title": "Linkedin Path",
            "type": "string"
          },
          "extraversion": {
            "title": "Extraversion level",
            "type": "integer",
            "default": 0
          },
          "neuroticism": {
            "title": "Neuroticism level",
            "type": "integer",
            "default": 0
          },
          "conscientiousness": {
            "title": "Conscientiousness level",
            "type": "integer",
            "default": 0
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "AuthBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}